Refactored to remove circular import problem.
Moved Trade-a-plane related classes into
`trade_a_plane` module.

Example
---------
 `crawler.ListingsPage` was moved to `trade_a_plane.ListingsPage`
